msm8996-common: Do not enable/disable msm-thermal hotplug
I'm not sure if it is because the blobs are updated or what,
or maybe some changes in the new kernel,
currently not only msm-thermal's hotplugging is not working,
it also causes CPU failure if thermal-engine is not running
and we enable/disable msm-thermal's core-control tunable.
Hence, removing this tunable line for now.

Initial inspection shows no evidence of msm-thermal's
hotplugging being disabled in kernel (it utilizes logic core count
to determine whether to enable core-control). Not going to
spend further time inspecting this since there are only two
use cases for now and they are working well:
1. thermal-engine enabled, we can mess with core-control.
2. thermal-engine disabled, we don't mess with core-control.